Unify sniffer constructor and move socket
