+++ pkg/transport/network/sniffer/options.go
@@ -1,0 +1,5 @@
+package sniffer
+
+type options interface {
+	Apply(*Sniffer) error
+}
