+++ pkg/transport/network/sniffer/sniffer.go
@@ -1,8 +1,7 @@
 package sniffer
 
 import (
-	"os"
-
+	"github.com/HyperloopUPV-H8/h9-backend/pkg/transport/network"
 	"github.com/google/gopacket"
 	"github.com/google/gopacket/layers"
 	"github.com/google/gopacket/pcap"
@@ -15,113 +14,51 @@ type Sniffer struct {
 	decoder *decoder
 }
 
-// NewOfflineSniffer creates a new sniffer that opens and reads from the specified file.
-// When firstLayer is set to nil the program tries to auto detect the link type for the connection, otherwise
-// the specified value is used.
-func NewOfflineSniffer(file string, firstLayer *gopacket.LayerType) (*Sniffer, error) {
-	source, err := pcap.OpenOffline(file)
-	if err != nil {
-		return nil, err
-	}
-
+// New creates a new sniffer from the provided source.
+//
+// source is a previously created pcap handle that will capture packets from the wire or a save file.
+//
+// firstLayer, when set to something, will be used as the firstLayer for the packet decoder. When this
+// value is nil, the program tries to automatically detect the first layer from the source.
+//
+// opts are a list of options to be applied to the sniffer used to configure the capture options.
+func New(source *pcap.Handle, firstLayer *gopacket.LayerType, opts ...options) (*Sniffer, error) {
 	first := source.LinkType().LayerType()
 	if firstLayer != nil {
 		first = *firstLayer
 	}
-
 	decoder := newDecoder(first)
 
-	return &Sniffer{
+	sniffer := &Sniffer{
 		source:  source,
 		decoder: &decoder,
-	}, nil
-}
-
-// NewOfflineFileSniffer creates a new sniffer that reads from the specified file.
-// When firstLayer is set to nil the program tries to auto detect the link type for the connection, otherwise
-// the specified value is used.
-func NewOfflineFileSniffer(file *os.File, firstLayer *gopacket.LayerType) (*Sniffer, error) {
-	source, err := pcap.OpenOfflineFile(file)
-	if err != nil {
-		return nil, err
-	}
-
-	first := source.LinkType().LayerType()
-	if firstLayer != nil {
-		first = *firstLayer
-	}
-
-	decoder := newDecoder(first)
-
-	return &Sniffer{
-		source:  source,
-		decoder: &decoder,
-	}, nil
-}
-
-type LiveSnifferOpt func(*pcap.InactiveHandle) error
-
-// NewLiveSniffer creates a new live sniffer capturing packets arriving on the specified device
-// When firstLayer is set to nil the program tries to auto detect the link type for the connection, otherwise
-// the specified value is used.
-// Multiple opts may also be provided, these will apply configuration options to the sniffer before
-// creating it.
-func NewLiveSniffer(device string, firstLayer *gopacket.LayerType, opts ...LiveSnifferOpt) (*Sniffer, error) {
-	handle, err := pcap.NewInactiveHandle(device)
-	defer handle.CleanUp()
-	if err != nil {
-		return nil, err
 	}
 
 	for _, opt := range opts {
-		if err = opt(handle); err != nil {
-			return nil, err
+		if err := opt.Apply(sniffer); err != nil {
+			return sniffer, err
 		}
 	}
 
-	source, err := handle.Activate()
-	if err != nil {
-		return nil, err
-	}
-
-	first := source.LinkType().LayerType()
-	if firstLayer != nil {
-		first = *firstLayer
-	}
-
-	decoder := newDecoder(first)
-
-	return &Sniffer{
-		source:  source,
-		decoder: &decoder,
-	}, nil
-}
-
-// Socket defines a unique conversation over a network by using the source and destination
-// IP addresses and TCP/UDP ports.
-type Socket struct {
-	SrcIP   string
-	SrcPort uint16
-	DstIP   string
-	DstPort uint16
+	return sniffer, nil
 }
 
 // ReadNext pulls the next packet from the wire, decodes it and returns the socket it belongs to,
 // its TCP or UDP payload and any errors encountered.
-func (sniffer *Sniffer) ReadNext() (Socket, []byte, error) {
+func (sniffer *Sniffer) ReadNext() (network.Socket, []byte, error) {
 	data, _, err := sniffer.source.ReadPacketData()
 	if err != nil {
-		return Socket{}, nil, err
+		return network.Socket{}, nil, err
 	}
 
 	packetLayers, err := sniffer.decoder.decode(data)
 	if err != nil {
-		return Socket{}, data, err
+		return network.Socket{}, data, err
 	}
 
 	ip := sniffer.decoder.IPv4()
 
-	socket := Socket{
+	socket := network.Socket{
 		SrcIP:   ip.SrcIP.String(),
 		SrcPort: 0,
 		DstIP:   ip.DstIP.String(),
@@ -145,7 +82,7 @@ layerLoop:
 	}
 
 	if socket.SrcPort == 0 && socket.DstPort == 0 {
-		return Socket{}, data, ErrMissingPayload{packetLayers}
+		return network.Socket{}, data, ErrMissingPayload{packetLayers}
 	}
 
 	return socket, sniffer.decoder.Payload(), nil
